Allow "better" development with docker-compose
We update the example_config.json in this commit so that out the box, the dockerized
version of the app can find the dockerized mysql server. In the future though we should
support env var permissions so we don't have to deal with config files at all, and can
ensure the correct config is hosted in the docker-compose.yml as injected ENV vars.