Add migration to ensure 'All Hosts' is complete
Due to recreating the 'All Hosts' label in #1282, we get inconsistent counts
for hosts that have not checked in since that migration. This seems acceptable
for other labels, but it is important that 'All Hosts' really includes all the
hosts.

This migration adds all the hosts into that label.

Fixes #1329