Set payload from argument ID in entity delete action
This fixes a bug in which the frontend expected the entity ID to be returned in
the response body of a deletion request. Because the API does not do this (and
we don't want to make it do this), the ID needs to be made available for
updating the UI after the request returns.

Fixes #1398