Rename remaining file in Admin to admin
We have files spread across both `Admin` and `admin`. On an osx machine, those are the same directory, which has probably masked some underlying errors. This normalizes the case to the one that has more.

I did not test this, it's possible we should make this all title case instead.